[Feat] Add SettingUI
