+++ ItaCH_Smash_Legends/Assets/Resources/UI/SettingPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0979436605755cb438c4ff62055634df
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
