+++ ItaCH_Smash_Legends/Assets/UI/Font/Exo2-ExtraBold SDF.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82573239799d86945a9d49ddc025627f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
