+++ ItaCH_Smash_Legends/Assets/UI/Script/ButtonScaler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ddc5c95a4e4d8774a8c972369180b286
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
