+++ ItaCH_Smash_Legends/Assets/UI/Script/VolumeSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1cf7a72314a6f4e48927d61252619b56
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
